prinf update - u and moove itoa

+++ srcs/ft_printf_i.c
@@ -6,13 +6,13 @@
 /*   By: cmaginot <cmaginot@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/14 14:23:09 by cmaginot          #+#    #+#             */
-/*   Updated: 2021/06/15 12:13:57 by cmaginot         ###   ########.fr       */
+/*   Updated: 2021/06/15 13:15:34 by cmaginot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_printf.h"
 
-int	ft_printf_i(t_value_printf *value_printf, int i)
+int	ft_printf_i(t_value_printf *value_printf, long long i)
 {
 	(void)value_printf;
